+++ creme/stats/kurtosis.py
@@ -77,7 +77,7 @@ class Kurtosis(moments.CentralMoments):
 
     References:
 
-    - `Welford's online algorithm <https://www.wikiwand.com/en/Algorithms_for_calculating_variance#/Welford's_Online_algorithm>`_
+    1. `Welford's online algorithm <https://www.wikiwand.com/en/Algorithms_for_calculating_variance#/Welford's_Online_algorithm>`_
 
     """