+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+- `Covariance` and `PearsonCorrelation` to `stats`
+
 ## [0.0.3](https://pypi.org/project/creme/0.0.3/) - 2019-03-21
 
 ### Added